refactor: move epicycle crud to own package
