+++ src/dearpywindow/window/__init__.py
@@ -1,8 +1,5 @@
 from .interface import IWindow
 from .main import MainWindow
 from .epicycle_menu import EpicyclesMenuWindow
-from .create_epicycle import CreateEpicycleWindow
-from .update_epicycle import UpdateEpicycleWindow
-from .delete_epicycle import DeleteEpicycleWindow
-from .create_child_epicycle import CreateChildEpicycleWindow
+from .epicycle_crud import CreateEpicycleWindow, UpdateEpicycleWindow, DeleteEpicycleWindow, CreateChildEpicycleWindow
 from . import popups
